Make ncpp::NCKey a struct
`NotCurses::getc` returns a `char32_t`, which is an open set of
characters/keycodes and so an `enum class` might not be the best choice
for `NCKey` here - it makes `switch` statements rather weird when they
have to `static_cast` the enum value to `char32_t`. `constexpr` values
work much better here. However, it is nice to have those values isolated
into a separate, struct, scope.